UPDATE installer and uninstaller

+++ installer.sh
@@ -1,19 +1,21 @@
 #!/bin/bash
 
-if type git > /dev/null 2>&1; then
-    git clone https://github.com/honos645/dotfiles
-elif type curl > /dev/null 2>&1; then
-    command curl -OL https://github.com/honos645/dotfiles/archive/refs/heads/main.tar.gz
-    command tar xvf main.tar.gz
-    command mv dotfiles-main dotfiles
-    command rm -rf main.tar.gz
-elif type wget > /dev/null 2>&1; then
-    command wget -O main.tar.gz https://github.com/honos645/dotfiles/archive/refs/heads/main.tar.gz
-    command tar xvf main.tar.gz
-    command mv dotfiles-main dotfiles
-    command rm -rf main.tar.gz
-else
-    echo "ERROR download failure"
+if [ ! -e  "${HOME}/.dotfiles" ]; then
+    if type git > /dev/null 2>&1; then
+        git clone https://github.com/honos645/dotfiles
+    elif type curl > /dev/null 2>&1; then
+        command curl -OL https://github.com/honos645/dotfiles/archive/refs/heads/main.tar.gz
+        command tar xvf main.tar.gz
+        command mv dotfiles-main dotfiles
+        command rm -rf main.tar.gz
+    elif type wget > /dev/null 2>&1; then
+        command wget -O main.tar.gz https://github.com/honos645/dotfiles/archive/refs/heads/main.tar.gz
+        command tar xvf main.tar.gz
+        command mv dotfiles-main dotfiles
+        command rm -rf main.tar.gz
+    else
+        echo "ERROR download failure"
+    fi
 fi
 
 command cd ./dotfiles
